FT module clean up.
* renaming features to: minting, burning, whitelisting, freezing etc
* logically sort feature
* replace FT with the Token
* remove lines related to starport scaffolding
* structure the FT events to follow the same design
* merge multiple logically grouped files
* allow issuer ignore the freezing
