+++ proto/coreum/asset/ft/v1/event.proto
@@ -3,12 +3,13 @@ package coreum.asset.ft.v1;
 
 import "gogoproto/gogo.proto";
 import "cosmos/base/v1beta1/coin.proto";
+
 import "coreum/asset/ft/v1/token.proto";
 
 option go_package = "github.com/CoreumFoundation/coreum/x/asset/ft/types";
 
-// EventTokenIssued is emitted on MsgIssueToken.
-message EventTokenIssued {
+// EventIssued is emitted on MsgIssue.
+message EventIssued {
   string denom = 1;
   string issuer = 2;
   string symbol = 3;
@@ -19,7 +20,7 @@ message EventTokenIssued {
     (gogoproto.nullable) = false
   ];
   string description = 7;
-  repeated TokenFeature features = 8;
+  repeated Feature features = 8;
   string burn_rate = 9 [
     (gogoproto.nullable) = false,
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Dec"
@@ -32,13 +33,20 @@ message EventTokenIssued {
 
 message EventFrozenAmountChanged {
   string account = 1;
-  cosmos.base.v1beta1.Coin previous_amount = 2 [(gogoproto.nullable) = false];
-  cosmos.base.v1beta1.Coin current_amount = 3 [(gogoproto.nullable) = false];
+  string denom = 2;
+  string previous_amount = 3 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
+    (gogoproto.nullable) = false
+  ];
+  string current_amount = 4 [
+    (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
+    (gogoproto.nullable) = false
+  ];
 }
 
 message EventWhitelistedAmountChanged {
   string account = 1;
-  string denom  = 2;
+  string denom = 2;
   string previous_amount = 3 [
     (gogoproto.customtype) = "github.com/cosmos/cosmos-sdk/types.Int",
     (gogoproto.nullable) = false
