+++ proto/coreum/asset/ft/v1/params.proto
@@ -1,15 +1,14 @@
 syntax = "proto3";
 package coreum.asset.ft.v1;
 
-import "cosmos/base/v1beta1/coin.proto";
-
 import "gogoproto/gogo.proto";
+import "cosmos/base/v1beta1/coin.proto";
 
 option go_package = "github.com/CoreumFoundation/coreum/x/asset/ft/types";
 
 // Params store gov manageable parameters.
 message Params {
-  // issue_fee is the fee burnt each time new FT is issued
+  // issue_fee is the fee burnt each time new token is issued
   cosmos.base.v1beta1.Coin issue_fee = 1 [
     (gogoproto.nullable) = false,
     (gogoproto.moretags) = "yaml:\"issue_fee\""
