+++ proto/coreum/asset/ft/v1/token.proto
@@ -6,21 +6,21 @@ import "cosmos/base/v1beta1/coin.proto";
 
 option go_package = "github.com/CoreumFoundation/coreum/x/asset/ft/types";
 
-// TokenFeature defines possible features of fungible token
-enum TokenFeature {
-  freeze = 0;
-  mint = 1;
-  burn = 2;
-  whitelist = 3;
+// Feature defines possible features of fungible token.
+enum Feature {
+  minting = 0;
+  burning = 1;
+  freezing = 2;
+  whitelisting = 3;
 }
 
-// FTDefinition defines the fungible token settings to store.
-message FTDefinition {
+// Definition defines the fungible token settings to store.
+message Definition {
   option (gogoproto.goproto_getters) = false;
   option (gogoproto.equal) = false;
   string denom = 1;
   string issuer = 2;
-  repeated TokenFeature features = 3;
+  repeated Feature features = 3;
   // burn_rate is a number between 0 and 1 which will be multiplied by send amount to determine
   // burn_amount. This value will be burnt on top of the send amount.
   string burn_rate = 4 [
@@ -35,8 +35,8 @@ message FTDefinition {
   ];
 }
 
-// FT is a full representation of the fungible token.
-message FT {
+// Token is a full representation of the fungible token.
+message Token {
   option (gogoproto.equal) = false;
   option (gogoproto.goproto_getters) = false;
   string denom = 1;
@@ -46,7 +46,7 @@ message FT {
   uint32 precision = 5;
   string description = 6;
   bool globally_frozen = 7;
-  repeated TokenFeature features = 8;
+  repeated Feature features = 8;
   // burn_rate is a number between 0 and 1 which will be multiplied by send amount to determine
   // burn_amount. This value will be burnt on top of the send amount.
   string burn_rate = 9 [
