+++ proto/coreum/asset/ft/v1/tx.proto
@@ -16,16 +16,16 @@ service Msg {
   // Issue defines a method to issue a new fungible token.
   rpc Issue(MsgIssue) returns (EmptyResponse);
 
-  // Mint mints new fungible tokens
+  // Mint mints new fungible tokens.
   rpc Mint(MsgMint) returns (EmptyResponse);
-  // Burn burns the specified fungible tokens from senders balance if the sender has enough balance
+  // Burn burns the specified fungible tokens from senders balance if the sender has enough balance.
   rpc Burn(MsgBurn) returns (EmptyResponse);
 
   // Freeze freezes a part of the fungible tokens in an
   // account, only if the freezable feature is enabled on that token.
   rpc Freeze(MsgFreeze) returns (EmptyResponse);
   // Unfreeze unfreezes a part of the frozen fungible tokens in an
-  // account, only if there are such frozen tokens on that account
+  // account, only if there are such frozen tokens on that account.
   rpc Unfreeze(MsgUnfreeze) returns (EmptyResponse);
 
   // GloballyFreeze freezes fungible token so no operations are allowed with it before unfrozen.
@@ -35,7 +35,7 @@ service Msg {
   // This operation is idempotent so global unfreezing of non-frozen token does nothing.
   rpc GloballyUnfreeze(MsgGloballyUnfreeze) returns (EmptyResponse);
 
-  // SetWhitelistedLimit sets the limit of how many tokens a specific account may hold
+  // SetWhitelistedLimit sets the limit of how many tokens a specific account may hold.
   rpc SetWhitelistedLimit(MsgSetWhitelistedLimit) returns (EmptyResponse);
 }
 
@@ -50,7 +50,7 @@ message MsgIssue {
     (gogoproto.nullable) = false
   ];
   string description = 6;
-  repeated TokenFeature features = 7;
+  repeated Feature features = 7;
   // burn_rate is a number between 0 and 1 which will be multiplied by send amount to determine
   // burn_amount. This value will be burnt on top of the send amount.
   string burn_rate = 8 [
@@ -65,6 +65,16 @@ message MsgIssue {
   ];
 }
 
+message MsgMint {
+  string sender = 1;
+  cosmos.base.v1beta1.Coin coin = 2 [(gogoproto.nullable) = false];
+}
+
+message MsgBurn {
+  string sender = 1;
+  cosmos.base.v1beta1.Coin coin = 3 [(gogoproto.nullable) = false];
+}
+
 message MsgFreeze {
   string sender = 1;
   string account = 2;
@@ -74,16 +84,6 @@ message MsgFreeze {
 message MsgUnfreeze {
   string sender = 1;
   string account = 2;
-  cosmos.base.v1beta1.Coin coin = 3 [(gogoproto.nullable) = false];
-}
-
-message MsgMint {
-  string sender = 1;
-  cosmos.base.v1beta1.Coin coin = 2 [(gogoproto.nullable) = false];
-}
-
-message MsgBurn {
-  string sender = 1;
   cosmos.base.v1beta1.Coin coin = 3 [(gogoproto.nullable) = false];
 }
