+++ x/asset/ft/keeper/grpc_query.go
@@ -14,12 +14,12 @@ var _ types.QueryServer = QueryService{}
 
 // QueryKeeper defines subscope of keeper methods required by query service.
 type QueryKeeper interface {
-	GetToken(ctx sdk.Context, denom string) (types.FT, error)
-	GetIssuerTokens(ctx sdk.Context, issuer sdk.AccAddress, pagination *query.PageRequest) ([]types.FT, *query.PageResponse, error)
-	GetFrozenBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin
+	GetIssuerTokens(ctx sdk.Context, issuer sdk.AccAddress, pagination *query.PageRequest) ([]types.Token, *query.PageResponse, error)
+	GetToken(ctx sdk.Context, denom string) (types.Token, error)
 	GetFrozenBalances(ctx sdk.Context, addr sdk.AccAddress, pagination *query.PageRequest) (sdk.Coins, *query.PageResponse, error)
-	GetWhitelistedBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin
+	GetFrozenBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin
 	GetWhitelistedBalances(ctx sdk.Context, addr sdk.AccAddress, pagination *query.PageRequest) (sdk.Coins, *query.PageResponse, error)
+	GetWhitelistedBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin
 }
 
 // QueryService serves grpc query requests for assets module.
@@ -32,18 +32,6 @@ func NewQueryService(keeper QueryKeeper) QueryService {
 	return QueryService{
 		keeper: keeper,
 	}
-}
-
-// Token queries an fungible token.
-func (qs QueryService) Token(ctx context.Context, req *types.QueryTokenRequest) (*types.QueryTokenResponse, error) {
-	token, err := qs.keeper.GetToken(sdk.UnwrapSDKContext(ctx), req.GetDenom())
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.QueryTokenResponse{
-		Token: token,
-	}, nil
 }
 
 // Tokens returns fungible tokens query result.
@@ -60,6 +48,18 @@ func (qs QueryService) Tokens(ctx context.Context, req *types.QueryTokensRequest
 	return &types.QueryTokensResponse{
 		Pagination: pageRes,
 		Tokens:     tokens,
+	}, nil
+}
+
+// Token queries an fungible token.
+func (qs QueryService) Token(ctx context.Context, req *types.QueryTokenRequest) (*types.QueryTokenResponse, error) {
+	token, err := qs.keeper.GetToken(sdk.UnwrapSDKContext(ctx), req.GetDenom())
+	if err != nil {
+		return nil, err
+	}
+
+	return &types.QueryTokenResponse{
+		Token: token,
 	}, nil
 }
