+++ x/asset/ft/keeper/keeper_test.go
@@ -3,7 +3,6 @@ package keeper_test
 import (
 	"errors"
 	"fmt"
-	"math/big"
 	"strings"
 	"testing"
 
@@ -21,128 +20,9 @@ import (
 	"github.com/CoreumFoundation/coreum/pkg/config/constant"
 	"github.com/CoreumFoundation/coreum/testutil/event"
 	"github.com/CoreumFoundation/coreum/testutil/simapp"
-	"github.com/CoreumFoundation/coreum/x/asset/ft/keeper"
 	"github.com/CoreumFoundation/coreum/x/asset/ft/types"
 	wbankkeeper "github.com/CoreumFoundation/coreum/x/wbank/keeper"
 )
-
-func TestKeeper_ValidateSymbol(t *testing.T) {
-	requireT := require.New(t)
-	testApp := simapp.New()
-	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
-	ftKeeper := testApp.AssetFTKeeper
-
-	unacceptableSymbols := []string{
-		"ABC/1",
-		"core",
-		"ucore",
-		"Core",
-		"uCore",
-		"CORE",
-		"UCORE",
-		"3abc",
-		"3ABC",
-	}
-
-	acceptableSymbols := []string{
-		"btc-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
-		"BTC-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
-		"ABC-1",
-		"ABC1",
-		"coreum",
-		"ucoreum",
-		"Coreum",
-		"uCoreum",
-		"COREeum",
-		"A1234567890123456789012345678901234567890123456789012345678901234567890",
-		"AB1234567890123456789012345678901234567890123456789012345678901234567890",
-	}
-
-	assertValidSymbol := func(symbol string, isValid bool) {
-		addr := sdk.AccAddress(ed25519.GenPrivKey().PubKey().Address())
-		settings := types.IssueSettings{
-			Issuer:        addr,
-			Symbol:        symbol,
-			Subunit:       "subunit",
-			Description:   "ABC Desc",
-			InitialAmount: sdk.NewInt(777),
-			Features:      []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
-		}
-
-		_, err := ftKeeper.Issue(ctx, settings)
-		if types.ErrInvalidInput.Is(err) == isValid {
-			requireT.Equal(types.ErrInvalidInput.Is(err), !isValid)
-		}
-	}
-
-	for _, symbol := range unacceptableSymbols {
-		assertValidSymbol(symbol, false)
-	}
-
-	for _, symbol := range acceptableSymbols {
-		assertValidSymbol(symbol, true)
-	}
-}
-
-func TestKeeper_ValidateSubunit(t *testing.T) {
-	requireT := require.New(t)
-	testApp := simapp.New()
-	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
-	ftKeeper := testApp.AssetFTKeeper
-
-	unacceptableSubunits := []string{
-		"ABC-1",
-		"ABC/1",
-		"btc-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
-		"BTC-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
-		"core",
-		"ucore",
-		"Core",
-		"uCore",
-		"CORE",
-		"UCORE",
-		"3abc",
-		"3ABC",
-		"uCoreum",
-		"Coreum",
-		"COREeum",
-		"AB1234567890123456789012345678901234567890123456789012345678901234567890",
-	}
-
-	acceptableSubunits := []string{
-		"abc1",
-		"coreum",
-		"ucoreum",
-		"a1234567890123456789012345678901234567890123456789012345678901234567890",
-	}
-
-	assertValidSubunit := func(subunit string, isValid bool) {
-		addr := sdk.AccAddress(ed25519.GenPrivKey().PubKey().Address())
-		settings := types.IssueSettings{
-			Issuer:        addr,
-			Symbol:        "symbol",
-			Subunit:       subunit,
-			Description:   "ABC Desc",
-			InitialAmount: sdk.NewInt(777),
-			Features:      []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
-		}
-
-		_, err := ftKeeper.Issue(ctx, settings)
-		if isValid {
-			requireT.NoError(err)
-		} else {
-			requireT.ErrorIs(types.ErrInvalidInput, err, "subunit", subunit)
-		}
-	}
-
-	for _, su := range unacceptableSubunits {
-		assertValidSubunit(su, false)
-	}
-
-	for _, su := range acceptableSubunits {
-		assertValidSubunit(su, true)
-	}
-}
 
 func TestKeeper_Issue(t *testing.T) {
 	requireT := require.New(t)
@@ -168,7 +48,7 @@ func TestKeeper_Issue(t *testing.T) {
 		Subunit:       "abc",
 		Precision:     8,
 		InitialAmount: sdk.NewInt(777),
-		Features:      []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
 	}
 
 	denom, err := ftKeeper.Issue(ctx, settings)
@@ -189,14 +69,14 @@ func TestKeeper_Issue(t *testing.T) {
 
 	gotToken, err := ftKeeper.GetToken(ctx, denom)
 	requireT.NoError(err)
-	requireT.Equal(types.FT{
+	requireT.Equal(types.Token{
 		Denom:              denom,
 		Issuer:             settings.Issuer.String(),
 		Symbol:             settings.Symbol,
 		Description:        settings.Description,
 		Subunit:            strings.ToLower(settings.Subunit),
 		Precision:          settings.Precision,
-		Features:           []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
+		Features:           []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
 		BurnRate:           sdk.NewDec(0),
 		SendCommissionRate: sdk.NewDec(0),
 	}, gotToken)
@@ -240,6 +120,124 @@ func TestKeeper_Issue(t *testing.T) {
 	requireT.True(errors.Is(types.ErrInvalidInput, err))
 }
 
+func TestKeeper_IssueValidateSymbol(t *testing.T) {
+	requireT := require.New(t)
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+	ftKeeper := testApp.AssetFTKeeper
+
+	unacceptableSymbols := []string{
+		"ABC/1",
+		"core",
+		"ucore",
+		"Core",
+		"uCore",
+		"CORE",
+		"UCORE",
+		"3abc",
+		"3ABC",
+	}
+
+	acceptableSymbols := []string{
+		"btc-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
+		"BTC-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
+		"ABC-1",
+		"ABC1",
+		"coreum",
+		"ucoreum",
+		"Coreum",
+		"uCoreum",
+		"COREeum",
+		"A1234567890123456789012345678901234567890123456789012345678901234567890",
+		"AB1234567890123456789012345678901234567890123456789012345678901234567890",
+	}
+
+	assertValidSymbol := func(symbol string, isValid bool) {
+		addr := sdk.AccAddress(ed25519.GenPrivKey().PubKey().Address())
+		settings := types.IssueSettings{
+			Issuer:        addr,
+			Symbol:        symbol,
+			Subunit:       "subunit",
+			Description:   "ABC Desc",
+			InitialAmount: sdk.NewInt(777),
+			Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
+		}
+
+		_, err := ftKeeper.Issue(ctx, settings)
+		if types.ErrInvalidInput.Is(err) == isValid {
+			requireT.Equal(types.ErrInvalidInput.Is(err), !isValid)
+		}
+	}
+
+	for _, symbol := range unacceptableSymbols {
+		assertValidSymbol(symbol, false)
+	}
+
+	for _, symbol := range acceptableSymbols {
+		assertValidSymbol(symbol, true)
+	}
+}
+
+func TestKeeper_IssueValidateSubunit(t *testing.T) {
+	requireT := require.New(t)
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+	ftKeeper := testApp.AssetFTKeeper
+
+	unacceptableSubunits := []string{
+		"ABC-1",
+		"ABC/1",
+		"btc-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
+		"BTC-devcore1phjrez5j2wp5qzp0zvlqavasvw60mkp2zmfe6h",
+		"core",
+		"ucore",
+		"Core",
+		"uCore",
+		"CORE",
+		"UCORE",
+		"3abc",
+		"3ABC",
+		"uCoreum",
+		"Coreum",
+		"COREeum",
+		"AB1234567890123456789012345678901234567890123456789012345678901234567890",
+	}
+
+	acceptableSubunits := []string{
+		"abc1",
+		"coreum",
+		"ucoreum",
+		"a1234567890123456789012345678901234567890123456789012345678901234567890",
+	}
+
+	assertValidSubunit := func(subunit string, isValid bool) {
+		addr := sdk.AccAddress(ed25519.GenPrivKey().PubKey().Address())
+		settings := types.IssueSettings{
+			Issuer:        addr,
+			Symbol:        "symbol",
+			Subunit:       subunit,
+			Description:   "ABC Desc",
+			InitialAmount: sdk.NewInt(777),
+			Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
+		}
+
+		_, err := ftKeeper.Issue(ctx, settings)
+		if isValid {
+			requireT.NoError(err)
+		} else {
+			requireT.ErrorIs(types.ErrInvalidInput, err, "subunit", subunit)
+		}
+	}
+
+	for _, su := range unacceptableSubunits {
+		assertValidSubunit(su, false)
+	}
+
+	for _, su := range acceptableSubunits {
+		assertValidSubunit(su, true)
+	}
+}
+
 func TestKeeper_Issue_WithZeroIssueFee(t *testing.T) {
 	requireT := require.New(t)
 
@@ -262,7 +260,7 @@ func TestKeeper_Issue_WithZeroIssueFee(t *testing.T) {
 		Subunit:       "abc",
 		Precision:     8,
 		InitialAmount: sdk.NewInt(777),
-		Features:      []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
 	}
 
 	_, err := ftKeeper.Issue(ctx, settings)
@@ -290,7 +288,7 @@ func TestKeeper_Issue_WithNoFundsCoveringFee(t *testing.T) {
 		Subunit:       "abc",
 		Precision:     8,
 		InitialAmount: sdk.NewInt(777),
-		Features:      []types.TokenFeature{types.TokenFeature_freeze}, //nolint:nosnakecase
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
 	}
 
 	_, err := ftKeeper.Issue(ctx, settings)
@@ -314,9 +312,9 @@ func TestKeeper_Mint(t *testing.T) {
 		Symbol:        "NotMintable",
 		Subunit:       "notmintable",
 		InitialAmount: sdk.NewInt(777),
-		Features: []types.TokenFeature{
-			types.TokenFeature_freeze, //nolint:nosnakecase
-			types.TokenFeature_burn,   //nolint:nosnakecase
+		Features: []types.Feature{
+			types.Feature_freezing, //nolint:nosnakecase
+			types.Feature_burning,  //nolint:nosnakecase
 		},
 	}
 
@@ -334,8 +332,8 @@ func TestKeeper_Mint(t *testing.T) {
 		Symbol:        "mintable",
 		Subunit:       "mintable",
 		InitialAmount: sdk.NewInt(777),
-		Features: []types.TokenFeature{
-			types.TokenFeature_mint, //nolint:nosnakecase
+		Features: []types.Feature{
+			types.Feature_minting, //nolint:nosnakecase
 		},
 	}
 
@@ -377,9 +375,9 @@ func TestKeeper_Burn(t *testing.T) {
 		Symbol:        "NotBurnable",
 		Subunit:       "notburnable",
 		InitialAmount: sdk.NewInt(777),
-		Features: []types.TokenFeature{
-			types.TokenFeature_freeze, //nolint:nosnakecase
-			types.TokenFeature_mint,   //nolint:nosnakecase
+		Features: []types.Feature{
+			types.Feature_freezing, //nolint:nosnakecase
+			types.Feature_minting,  //nolint:nosnakecase
 		},
 	}
 
@@ -405,9 +403,9 @@ func TestKeeper_Burn(t *testing.T) {
 		Symbol:        "burnable",
 		Subunit:       "burnable",
 		InitialAmount: sdk.NewInt(777),
-		Features: []types.TokenFeature{
-			types.TokenFeature_burn,   //nolint:nosnakecase
-			types.TokenFeature_freeze, //nolint:nosnakecase
+		Features: []types.Feature{
+			types.Feature_burning,  //nolint:nosnakecase
+			types.Feature_freezing, //nolint:nosnakecase
 		},
 	}
 
@@ -415,7 +413,7 @@ func TestKeeper_Burn(t *testing.T) {
 	requireT.NoError(err)
 
 	// send to new recipient address
-	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(sdk.NewCoin(burnableDenom, sdk.NewInt(100))))
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(sdk.NewCoin(burnableDenom, sdk.NewInt(200))))
 	requireT.NoError(err)
 
 	// try to burn as non-issuer
@@ -427,240 +425,104 @@ func TestKeeper_Burn(t *testing.T) {
 	requireT.NoError(err)
 
 	balance := bankKeeper.GetBalance(ctx, issuer, burnableDenom)
-	requireT.EqualValues(sdk.NewCoin(burnableDenom, sdk.NewInt(577)), balance)
+	requireT.EqualValues(sdk.NewCoin(burnableDenom, sdk.NewInt(477)), balance)
 
 	totalSupply, err := bankKeeper.TotalSupply(sdk.WrapSDKContext(ctx), &banktypes.QueryTotalSupplyRequest{})
 	requireT.NoError(err)
 	requireT.EqualValues(sdk.NewInt(577), totalSupply.Supply.AmountOf(burnableDenom))
 
-	// try to burn frozen amount
+	// try to burn issuer frozen amount
 	err = ftKeeper.Freeze(ctx, issuer, issuer, sdk.NewCoin(burnableDenom, sdk.NewInt(600)))
 	requireT.NoError(err)
-
+	// issuer can burn even frozen
 	err = ftKeeper.Burn(ctx, issuer, sdk.NewCoin(burnableDenom, sdk.NewInt(100)))
+	requireT.NoError(err)
+
+	// try to burn non-issuer frozen coins
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(burnableDenom, sdk.NewInt(100)))
+	requireT.NoError(err)
+	err = ftKeeper.Burn(ctx, recipient, sdk.NewCoin(burnableDenom, sdk.NewInt(100)))
 	requireT.ErrorIs(sdkerrors.ErrInsufficientFunds, err)
 }
 
-//nolint:funlen // there are too many tests cases
-func TestKeeperCalculateRateShare(t *testing.T) {
-	genAccount := func() string {
-		return sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address()).String()
-	}
-	var accounts []string
-	for i := 0; i < 11; i++ {
-		accounts = append(accounts, genAccount())
-	}
-	issuer := genAccount()
-	pow10 := func(ex int64) sdk.Int {
-		return sdk.NewIntFromBigInt(big.NewInt(0).Exp(big.NewInt(10), big.NewInt(ex), nil))
-	}
-	testCases := []struct {
-		rate      string
-		senders   map[string]sdk.Int
-		receivers map[string]sdk.Int
-		shares    map[string]sdk.Int
-	}{
-		{
-			rate:    "0.5",
-			senders: map[string]sdk.Int{},
-			shares:  map[string]sdk.Int{},
-		},
-		{
-			rate: "0.5",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(5),
-				accounts[1]: sdk.NewInt(5),
-			},
-			receivers: map[string]sdk.Int{
-				issuer: sdk.NewInt(10),
-			},
-			shares: map[string]sdk.Int{},
-		},
-		{
-			rate: "0.5",
-			senders: map[string]sdk.Int{
-				issuer: sdk.NewInt(10),
-			},
-			receivers: map[string]sdk.Int{
-				accounts[5]: sdk.NewInt(5),
-				accounts[6]: sdk.NewInt(5),
-			},
-			shares: map[string]sdk.Int{},
-		},
-		{
-			rate: "0.1",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(400),
-				accounts[1]: sdk.NewInt(600),
-			},
-			receivers: map[string]sdk.Int{
-				accounts[10]: sdk.NewInt(1000),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(40),
-				accounts[1]: sdk.NewInt(60),
-			},
-		},
-		{
-			rate: "0.1",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(399),
-				accounts[1]: sdk.NewInt(602),
-			},
-			receivers: map[string]sdk.Int{
-				accounts[10]: sdk.NewInt(1001),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(40),
-				accounts[1]: sdk.NewInt(61),
-			},
-		},
-		{
-			rate: "0.1",
-			senders: map[string]sdk.Int{
-				issuer:      sdk.NewInt(90),
-				accounts[0]: sdk.NewInt(29),
-				accounts[1]: sdk.NewInt(32),
-			},
-			receivers: map[string]sdk.Int{
-				genAccount(): sdk.NewInt(90 + 29 + 32),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(3),
-				accounts[1]: sdk.NewInt(4),
-			},
-		},
-		{
-			rate: "0.01",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(30000),
-				accounts[1]: sdk.NewInt(20000),
-			},
-			receivers: map[string]sdk.Int{
-				issuer:       sdk.NewInt(30000),
-				genAccount(): sdk.NewInt(20000),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(120),
-				accounts[1]: sdk.NewInt(80),
-			},
-		},
-		{
-			rate: "0.01001",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(30000),
-				accounts[1]: sdk.NewInt(20000),
-			},
-			receivers: map[string]sdk.Int{
-				issuer:       sdk.NewInt(30000),
-				genAccount(): sdk.NewInt(20000),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(121),
-				accounts[1]: sdk.NewInt(81),
-			},
-		},
-		{
-			rate: "0.1234",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(80),
-				accounts[1]: sdk.NewInt(17),
-			},
-			receivers: map[string]sdk.Int{
-				genAccount(): sdk.NewInt(97),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(10),
-				accounts[1]: sdk.NewInt(3),
-			},
-		},
-		{
-			rate: "0.1",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(1),
-				accounts[1]: sdk.NewInt(2),
-				accounts[2]: sdk.NewInt(9),
-			},
-			receivers: map[string]sdk.Int{
-				genAccount(): sdk.NewInt(12),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(1),
-				accounts[1]: sdk.NewInt(1),
-				accounts[2]: sdk.NewInt(1),
-			},
-		},
-		{
-			rate: "0.01",
-			senders: map[string]sdk.Int{
-				issuer:      sdk.NewInt(2100),
-				accounts[0]: sdk.NewInt(1100),
-				accounts[1]: sdk.NewInt(1700),
-				accounts[2]: sdk.NewInt(1900),
-			},
-			receivers: map[string]sdk.Int{
-				issuer:       sdk.NewInt(2100),
-				genAccount(): sdk.NewInt(300),
-				genAccount(): sdk.NewInt(1100),
-				genAccount(): sdk.NewInt(3300),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(11),
-				accounts[1]: sdk.NewInt(17),
-				accounts[2]: sdk.NewInt(19),
-			},
-		},
-		{
-			rate: "0.01",
-			senders: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(100).Mul(pow10(24)),
-				accounts[1]: sdk.NewInt(300).Mul(pow10(25)),
-				accounts[2]: sdk.NewInt(500).Mul(pow10(26)),
-			},
-			receivers: map[string]sdk.Int{
-				genAccount(): sdk.NewInt(100).Mul(pow10(24)),
-				genAccount(): sdk.NewInt(300).Mul(pow10(25)),
-				genAccount(): sdk.NewInt(500).Mul(pow10(26)),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(100).Mul(pow10(22)),
-				accounts[1]: sdk.NewInt(300).Mul(pow10(23)),
-				accounts[2]: sdk.NewInt(500).Mul(pow10(24)),
-			},
-		},
-		{
-			rate: "0.99",
-			senders: map[string]sdk.Int{
-				issuer:      sdk.NewInt(2100),
-				accounts[0]: sdk.NewInt(1100),
-				accounts[1]: sdk.NewInt(1700),
-				accounts[2]: sdk.NewInt(2728),
-			},
-			receivers: map[string]sdk.Int{
-				genAccount(): sdk.NewInt(2100),
-				genAccount(): sdk.NewInt(1000),
-				genAccount(): sdk.NewInt(1800),
-				genAccount(): sdk.NewInt(2728),
-			},
-			shares: map[string]sdk.Int{
-				accounts[0]: sdk.NewInt(1089),
-				accounts[1]: sdk.NewInt(1683),
-				accounts[2]: sdk.NewInt(2701),
-			},
-		},
+//nolint:dupl // We don't care
+func TestKeeper_BurnRate_BankSend(t *testing.T) {
+	requireT := require.New(t)
+
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+
+	assetKeeper := testApp.AssetFTKeeper
+	bankKeeper := testApp.BankKeeper
+	ba := newBankAsserter(ctx, t, bankKeeper)
+
+	// issue with more than 1 burn rate
+	issuer := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	settings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Precision:     6,
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(600),
+		Features:      []types.Feature{},
+		BurnRate:      sdk.MustNewDecFromStr("1.01"),
 	}
 
-	for _, tc := range testCases {
-		tc := tc
-		name := fmt.Sprintf("%+v", tc)
-		t.Run(name, func(t *testing.T) {
-			assertT := assert.New(t)
-			shares := keeper.CalculateRateShares(sdk.MustNewDecFromStr(tc.rate), issuer, tc.senders, tc.receivers)
-			for account, share := range shares {
-				assertT.EqualValues(tc.shares[account].String(), share.String())
-			}
-		})
+	_, err := assetKeeper.Issue(ctx, settings)
+	requireT.ErrorIs(types.ErrInvalidInput, err)
+
+	// issue token
+	settings = types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Precision:     6,
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(600),
+		Features:      []types.Feature{},
+		BurnRate:      sdk.MustNewDecFromStr("0.25"),
 	}
+
+	denom, err := assetKeeper.Issue(ctx, settings)
+	requireT.NoError(err)
+
+	recipient := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	// send from issuer to recipient (burn must not apply)
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(
+		sdk.NewCoin(denom, sdk.NewInt(500)),
+	))
+	requireT.NoError(err)
+
+	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
+		&recipient: 500,
+		&issuer:    100,
+	})
+
+	// send from recipient1 to recipient2 (burn must apply)
+	recipient2 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = bankKeeper.SendCoins(ctx, recipient, recipient2, sdk.NewCoins(
+		sdk.NewCoin(denom, sdk.NewInt(100)),
+	))
+	requireT.NoError(err)
+
+	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
+		&recipient:  375,
+		&recipient2: 100,
+		&issuer:     100,
+	})
+
+	// send from recipient to issuer account (burn must not apply)
+	err = bankKeeper.SendCoins(ctx, recipient, issuer, sdk.NewCoins(
+		sdk.NewCoin(denom, sdk.NewInt(375)),
+	))
+	requireT.NoError(err)
+
+	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
+		&recipient2: 100,
+		&issuer:     475,
+	})
 }
 
 //nolint:funlen // This is a complex test scenario and breaking it down will make it harder to read
@@ -687,7 +549,7 @@ func TestKeeper_BurnRate_BankMultiSend(t *testing.T) {
 			Precision:     6,
 			Description:   "DEF Desc",
 			InitialAmount: sdk.NewInt(1000),
-			Features:      []types.TokenFeature{},
+			Features:      []types.Feature{},
 			BurnRate:      sdk.MustNewDecFromStr(fmt.Sprintf("0.%d", i+1)),
 		}
 
@@ -700,7 +562,7 @@ func TestKeeper_BurnRate_BankMultiSend(t *testing.T) {
 		recipients = append(recipients, sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address()))
 	}
 
-	var testCases = []struct {
+	testCases := []struct {
 		name         string
 		inputs       []banktypes.Input
 		outputs      []banktypes.Output
@@ -834,86 +696,6 @@ func TestKeeper_BurnRate_BankMultiSend(t *testing.T) {
 }
 
 //nolint:dupl // We don't care
-func TestKeeper_BurnRate_BankSend(t *testing.T) {
-	requireT := require.New(t)
-
-	testApp := simapp.New()
-	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
-
-	assetKeeper := testApp.AssetFTKeeper
-	bankKeeper := testApp.BankKeeper
-	ba := newBankAsserter(ctx, t, bankKeeper)
-
-	// issue with more than 1 burn rate
-	issuer := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
-	settings := types.IssueSettings{
-		Issuer:        issuer,
-		Symbol:        "DEF",
-		Subunit:       "def",
-		Precision:     6,
-		Description:   "DEF Desc",
-		InitialAmount: sdk.NewInt(600),
-		Features:      []types.TokenFeature{},
-		BurnRate:      sdk.MustNewDecFromStr("1.01"),
-	}
-
-	_, err := assetKeeper.Issue(ctx, settings)
-	requireT.ErrorIs(types.ErrInvalidInput, err)
-
-	// issue token
-	settings = types.IssueSettings{
-		Issuer:        issuer,
-		Symbol:        "DEF",
-		Subunit:       "def",
-		Precision:     6,
-		Description:   "DEF Desc",
-		InitialAmount: sdk.NewInt(600),
-		Features:      []types.TokenFeature{},
-		BurnRate:      sdk.MustNewDecFromStr("0.25"),
-	}
-
-	denom, err := assetKeeper.Issue(ctx, settings)
-	requireT.NoError(err)
-
-	recipient := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
-
-	// send from issuer to recipient (burn must not apply)
-	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(
-		sdk.NewCoin(denom, sdk.NewInt(500)),
-	))
-	requireT.NoError(err)
-
-	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
-		&recipient: 500,
-		&issuer:    100,
-	})
-
-	// send from recipient1 to recipient2 (burn must apply)
-	recipient2 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
-	err = bankKeeper.SendCoins(ctx, recipient, recipient2, sdk.NewCoins(
-		sdk.NewCoin(denom, sdk.NewInt(100)),
-	))
-	requireT.NoError(err)
-
-	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
-		&recipient:  375,
-		&recipient2: 100,
-		&issuer:     100,
-	})
-
-	// send from recipient to issuer account (burn must not apply)
-	err = bankKeeper.SendCoins(ctx, recipient, issuer, sdk.NewCoins(
-		sdk.NewCoin(denom, sdk.NewInt(375)),
-	))
-	requireT.NoError(err)
-
-	ba.assertCoinDistribution(denom, map[*sdk.AccAddress]int64{
-		&recipient2: 100,
-		&issuer:     475,
-	})
-}
-
-//nolint:dupl // We don't care
 func TestKeeper_SendCommissionRate_BankSend(t *testing.T) {
 	requireT := require.New(t)
 
@@ -933,7 +715,7 @@ func TestKeeper_SendCommissionRate_BankSend(t *testing.T) {
 		Precision:          6,
 		Description:        "DEF Desc",
 		InitialAmount:      sdk.NewInt(600),
-		Features:           []types.TokenFeature{},
+		Features:           []types.Feature{},
 		SendCommissionRate: sdk.MustNewDecFromStr("1.01"),
 	}
 
@@ -948,7 +730,7 @@ func TestKeeper_SendCommissionRate_BankSend(t *testing.T) {
 		Precision:          6,
 		Description:        "DEF Desc",
 		InitialAmount:      sdk.NewInt(600),
-		Features:           []types.TokenFeature{},
+		Features:           []types.Feature{},
 		SendCommissionRate: sdk.MustNewDecFromStr("0.25"),
 	}
 
@@ -1012,7 +794,7 @@ func TestKeeper_BurnRateAndSendCommissionRate_BankSend(t *testing.T) {
 		Precision:          6,
 		Description:        "DEF Desc",
 		InitialAmount:      sdk.NewInt(600),
-		Features:           []types.TokenFeature{},
+		Features:           []types.Feature{},
 		BurnRate:           sdk.MustNewDecFromStr("0.5"),
 		SendCommissionRate: sdk.MustNewDecFromStr("0.25"),
 	}
@@ -1058,17 +840,503 @@ func TestKeeper_BurnRateAndSendCommissionRate_BankSend(t *testing.T) {
 	})
 }
 
-func (ba bankAssertion) assertCoinDistribution(denom string, dist map[*sdk.AccAddress]int64) {
-	requireT := require.New(ba.t)
-	total := int64(0)
-	for acc, expectedBalance := range dist {
-		total += expectedBalance
-		getBalance := ba.bk.GetBalance(ba.ctx, *acc, denom)
-		requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(expectedBalance)).String(), getBalance.String())
+//nolint:funlen // this is complex test scenario and breaking it down is not helpful
+func TestKeeper_FreezeUnfreeze(t *testing.T) {
+	requireT := require.New(t)
+	assertT := assert.New(t)
+
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+
+	ftKeeper := testApp.AssetFTKeeper
+	bankKeeper := testApp.BankKeeper
+
+	issuer := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	settings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(666),
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
 	}
 
-	totalSupply := ba.bk.GetSupply(ba.ctx, denom)
-	requireT.Equal(totalSupply.String(), sdk.NewCoin(denom, sdk.NewInt(total)).String())
+	denom, err := ftKeeper.Issue(ctx, settings)
+	requireT.NoError(err)
+
+	unfreezableSettings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "ABC",
+		Subunit:       "abc",
+		Description:   "ABC Desc",
+		InitialAmount: sdk.NewInt(666),
+		Features:      []types.Feature{},
+	}
+
+	unfreezableDenom, err := ftKeeper.Issue(ctx, unfreezableSettings)
+	requireT.NoError(err)
+	_, err = ftKeeper.GetToken(ctx, unfreezableDenom)
+	requireT.NoError(err)
+
+	recipient := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(
+		sdk.NewCoin(denom, sdk.NewInt(100)),
+		sdk.NewCoin(unfreezableDenom, sdk.NewInt(100)),
+	))
+	requireT.NoError(err)
+
+	// try to freeze non-existent denom
+	nonExistentDenom := types.BuildDenom("nonexist", issuer)
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(nonExistentDenom, sdk.NewInt(10)))
+	assertT.True(sdkerrors.IsOf(err, types.ErrTokenNotFound))
+
+	// try to freeze unfreezable Token
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(unfreezableDenom, sdk.NewInt(10)))
+	assertT.ErrorIs(types.ErrFeatureDisabled, err)
+
+	// try to freeze from non issuer address
+	randomAddr := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = ftKeeper.Freeze(ctx, randomAddr, recipient, sdk.NewCoin(denom, sdk.NewInt(10)))
+	assertT.ErrorIs(sdkerrors.ErrUnauthorized, err)
+
+	// try to freeze 0 balance
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(0)))
+	requireT.ErrorIs(sdkerrors.ErrInvalidCoins, err)
+
+	// try to unfreeze 0 balance
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(0)))
+	requireT.ErrorIs(sdkerrors.ErrInvalidCoins, err)
+
+	// try to freeze more than balance
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(110)))
+	requireT.NoError(err)
+	frozenBalance := ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	assertT.EqualValues(sdk.NewCoin(denom, sdk.NewInt(110)), frozenBalance)
+
+	// try to unfreeze more than frozen balance
+	err = ftKeeper.Unfreeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(130)))
+	requireT.ErrorIs(sdkerrors.ErrInsufficientFunds, err)
+	frozenBalance = ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	assertT.EqualValues(sdk.NewCoin(denom, sdk.NewInt(110)), frozenBalance)
+
+	// set frozen balance back to zero
+	err = ftKeeper.Unfreeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(110)))
+	requireT.NoError(err)
+	frozenBalance = ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	assertT.EqualValues(sdk.NewCoin(denom, sdk.NewInt(0)).String(), frozenBalance.String())
+
+	// freeze, query frozen
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(40)))
+	requireT.NoError(err)
+	frozenBalance = ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(40)).String(), frozenBalance.String())
+
+	// test query all frozen
+	allBalances, pageRes, err := ftKeeper.GetAccountsFrozenBalances(ctx, &query.PageRequest{})
+	assertT.NoError(err)
+	assertT.Len(allBalances, 1)
+	assertT.EqualValues(1, pageRes.GetTotal())
+	assertT.EqualValues(recipient.String(), allBalances[0].Address)
+	requireT.Equal(sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(40))).String(), allBalances[0].Coins.String())
+
+	// increase frozen and query
+	err = ftKeeper.Freeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(40)))
+	requireT.NoError(err)
+	frozenBalance = ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(80)), frozenBalance)
+
+	// try to send more than available
+	coinsToSend := sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(80)))
+	// send
+	err = bankKeeper.SendCoins(ctx, recipient, issuer, coinsToSend)
+	assertT.True(sdkerrors.IsOf(err, sdkerrors.ErrInsufficientFunds))
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: recipient.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: issuer.String(), Coins: coinsToSend}})
+	assertT.True(sdkerrors.IsOf(err, sdkerrors.ErrInsufficientFunds))
+
+	// try to send unfrozen balance
+	recipient2 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	coinsToSend = sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(10)))
+	// send
+	err = bankKeeper.SendCoins(ctx, recipient, recipient2, coinsToSend)
+	requireT.NoError(err)
+	balance := bankKeeper.GetBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(90)), balance)
+	balance = bankKeeper.GetBalance(ctx, recipient2, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(10)), balance)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: recipient.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: recipient2.String(), Coins: coinsToSend}})
+	requireT.NoError(err)
+	balance = bankKeeper.GetBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(80)), balance)
+	balance = bankKeeper.GetBalance(ctx, recipient2, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(20)), balance)
+
+	// try to unfreeze from non issuer address
+	err = ftKeeper.Unfreeze(ctx, randomAddr, recipient, sdk.NewCoin(denom, sdk.NewInt(80)))
+	assertT.True(sdkerrors.IsOf(err, sdkerrors.ErrUnauthorized))
+
+	// unfreeze, query frozen, and try to send
+	err = ftKeeper.Unfreeze(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(80)))
+	requireT.NoError(err)
+	frozenBalance = ftKeeper.GetFrozenBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(0)), frozenBalance)
+	coinsToSend = sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(40)))
+	// send
+	err = bankKeeper.SendCoins(ctx, recipient, recipient2, coinsToSend)
+	requireT.NoError(err)
+	balance = bankKeeper.GetBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(40)), balance)
+	balance = bankKeeper.GetBalance(ctx, recipient2, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(60)), balance)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: recipient.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: recipient2.String(), Coins: coinsToSend}})
+	requireT.NoError(err)
+	balance = bankKeeper.GetBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(0)), balance)
+	balance = bankKeeper.GetBalance(ctx, recipient2, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(100)), balance)
+}
+
+//nolint:funlen // this is complex tests scenario and breaking it down is not beneficial
+func TestKeeper_GlobalFreezeUnfreeze(t *testing.T) {
+	requireT := require.New(t)
+	assertT := assert.New(t)
+
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+
+	ftKeeper := testApp.AssetFTKeeper
+	bankKeeper := testApp.BankKeeper
+
+	issuer := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	freezableSettings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "FREEZE",
+		Subunit:       "freeze",
+		Precision:     6,
+		Description:   "FREEZE Desc",
+		InitialAmount: sdk.NewInt(777),
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
+	}
+
+	freezableDenom, err := ftKeeper.Issue(ctx, freezableSettings)
+	requireT.NoError(err)
+
+	unfreezableSettings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "NOFREEZE",
+		Subunit:       "nofreeze",
+		Precision:     6,
+		Description:   "NOFREEZE Desc",
+		InitialAmount: sdk.NewInt(777),
+		Features:      []types.Feature{},
+	}
+
+	unfreezableDenom, err := ftKeeper.Issue(ctx, unfreezableSettings)
+	requireT.NoError(err)
+	_, err = ftKeeper.GetToken(ctx, unfreezableDenom)
+	requireT.NoError(err)
+
+	recipient := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, sdk.NewCoins(
+		sdk.NewCoin(freezableDenom, sdk.NewInt(100)),
+		sdk.NewCoin(unfreezableDenom, sdk.NewInt(100)),
+	))
+	requireT.NoError(err)
+
+	// try to global-freeze non-existent
+	nonExistentDenom := types.BuildDenom("nonexist", issuer)
+	err = ftKeeper.GloballyFreeze(ctx, issuer, nonExistentDenom)
+	assertT.True(sdkerrors.IsOf(err, types.ErrTokenNotFound))
+
+	// try to global-freeze unfreezable Token
+	err = ftKeeper.GloballyFreeze(ctx, issuer, unfreezableDenom)
+	assertT.ErrorIs(types.ErrFeatureDisabled, err)
+
+	// try to global-freeze from non issuer address
+	randomAddr := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = ftKeeper.GloballyFreeze(ctx, randomAddr, freezableDenom)
+	assertT.ErrorIs(sdkerrors.ErrUnauthorized, err)
+
+	// freeze twice to check global-freeze idempotence
+	err = ftKeeper.GloballyFreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	err = ftKeeper.GloballyFreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	frozenFt, err := ftKeeper.GetToken(ctx, freezableDenom)
+	requireT.NoError(err)
+	assertT.True(frozenFt.GloballyFrozen)
+
+	// try to global-unfreeze from non issuer address
+	err = ftKeeper.GloballyUnfreeze(ctx, randomAddr, freezableDenom)
+	assertT.ErrorIs(sdkerrors.ErrUnauthorized, err)
+
+	// unfreeze twice to check global-unfreeze idempotence
+	err = ftKeeper.GloballyUnfreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	err = ftKeeper.GloballyUnfreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	unfrozenFt, err := ftKeeper.GetToken(ctx, freezableDenom)
+	requireT.NoError(err)
+	assertT.False(unfrozenFt.GloballyFrozen)
+
+	// freeze, try to send & verify balance
+	err = ftKeeper.GloballyFreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	coinsToSend := sdk.NewCoins(sdk.NewCoin(freezableDenom, sdk.NewInt(10)))
+	// send
+	err = bankKeeper.SendCoins(ctx, recipient, randomAddr, coinsToSend)
+	assertT.ErrorIs(types.ErrGloballyFrozen, err)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: recipient.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: randomAddr.String(), Coins: coinsToSend}})
+	assertT.ErrorIs(types.ErrGloballyFrozen, err)
+
+	// unfreeze, try to send & verify balance
+	err = ftKeeper.GloballyUnfreeze(ctx, issuer, freezableDenom)
+	requireT.NoError(err)
+	coinsToSend = sdk.NewCoins(sdk.NewCoin(freezableDenom, sdk.NewInt(6)))
+	// send
+	err = bankKeeper.SendCoins(ctx, recipient, randomAddr, coinsToSend)
+	requireT.NoError(err)
+	balance := bankKeeper.GetBalance(ctx, randomAddr, freezableDenom)
+	requireT.Equal(sdk.NewCoin(freezableDenom, sdk.NewInt(6)), balance)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: recipient.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: randomAddr.String(), Coins: coinsToSend}})
+	requireT.NoError(err)
+	balance = bankKeeper.GetBalance(ctx, randomAddr, freezableDenom)
+	requireT.Equal(sdk.NewCoin(freezableDenom, sdk.NewInt(12)), balance)
+}
+
+//nolint:funlen // this is complex test scenario and breaking it down is not helpful
+func TestKeeper_Whitelist(t *testing.T) {
+	requireT := require.New(t)
+	assertT := assert.New(t)
+
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+
+	ftKeeper := testApp.AssetFTKeeper
+	bankKeeper := testApp.BankKeeper
+
+	issuer := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	settings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(666),
+		Features:      []types.Feature{types.Feature_whitelisting}, //nolint:nosnakecase
+	}
+
+	denom, err := ftKeeper.Issue(ctx, settings)
+	requireT.NoError(err)
+
+	unwhitelistableSettings := types.IssueSettings{
+		Issuer:        issuer,
+		Symbol:        "ABC",
+		Subunit:       "abc",
+		Description:   "ABC Desc",
+		InitialAmount: sdk.NewInt(666),
+		Features:      []types.Feature{},
+	}
+
+	recipient := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	unwhitelistableDenom, err := ftKeeper.Issue(ctx, unwhitelistableSettings)
+	requireT.NoError(err)
+	_, err = ftKeeper.GetToken(ctx, unwhitelistableDenom)
+	requireT.NoError(err)
+
+	// whitelisting fails on unwhitelistable token
+	err = ftKeeper.SetWhitelistedBalance(ctx, issuer, recipient, sdk.NewCoin(unwhitelistableDenom, sdk.NewInt(1)))
+	requireT.ErrorIs(types.ErrFeatureDisabled, err)
+
+	// try to whitelist non-existent denom
+	nonExistentDenom := types.BuildDenom("nonexist", issuer)
+	err = ftKeeper.SetWhitelistedBalance(ctx, issuer, recipient, sdk.NewCoin(nonExistentDenom, sdk.NewInt(10)))
+	assertT.True(sdkerrors.IsOf(err, types.ErrTokenNotFound))
+
+	// try to whitelist from non issuer address
+	randomAddr := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	err = ftKeeper.SetWhitelistedBalance(ctx, randomAddr, recipient, sdk.NewCoin(denom, sdk.NewInt(10)))
+	assertT.ErrorIs(sdkerrors.ErrUnauthorized, err)
+
+	// set whitelisted balance to 0
+	requireT.NoError(ftKeeper.SetWhitelistedBalance(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(0))))
+	whitelistedBalance := ftKeeper.GetWhitelistedBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(0)).String(), whitelistedBalance.String())
+
+	coinsToSend := sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(100)))
+	// send
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, coinsToSend)
+	requireT.ErrorIs(types.ErrWhitelistedLimitExceeded, err)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: issuer.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: recipient.String(), Coins: coinsToSend}})
+	requireT.True(types.ErrWhitelistedLimitExceeded.Is(err))
+
+	// set whitelisted balance to 100
+	requireT.NoError(ftKeeper.SetWhitelistedBalance(ctx, issuer, recipient, sdk.NewCoin(denom, sdk.NewInt(100))))
+	whitelistedBalance = ftKeeper.GetWhitelistedBalance(ctx, recipient, denom)
+	requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(100)).String(), whitelistedBalance.String())
+
+	// test query all whitelisted balances
+	allBalances, pageRes, err := ftKeeper.GetAccountsWhitelistedBalances(ctx, &query.PageRequest{})
+	assertT.NoError(err)
+	assertT.Len(allBalances, 1)
+	assertT.EqualValues(1, pageRes.GetTotal())
+	assertT.EqualValues(recipient.String(), allBalances[0].Address)
+	requireT.Equal(sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(100))).String(), allBalances[0].Coins.String())
+
+	coinsToSend = sdk.NewCoins(
+		sdk.NewCoin(denom, sdk.NewInt(50)),
+		sdk.NewCoin(unwhitelistableDenom, sdk.NewInt(50)),
+	)
+	// send
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, coinsToSend)
+	requireT.NoError(err)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: issuer.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: recipient.String(), Coins: coinsToSend}})
+	requireT.NoError(err)
+
+	// try to send more
+	coinsToSend = sdk.NewCoins(sdk.NewCoin(denom, sdk.NewInt(1)))
+	// send
+	err = bankKeeper.SendCoins(ctx, issuer, recipient, coinsToSend)
+	requireT.ErrorIs(types.ErrWhitelistedLimitExceeded, err)
+	// multi-send
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{{Address: issuer.String(), Coins: coinsToSend}},
+		[]banktypes.Output{{Address: recipient.String(), Coins: coinsToSend}})
+	requireT.ErrorIs(types.ErrWhitelistedLimitExceeded, err)
+
+	// try to whitelist from non issuer address
+	err = ftKeeper.SetWhitelistedBalance(ctx, randomAddr, recipient, sdk.NewCoin(denom, sdk.NewInt(80)))
+	assertT.True(sdkerrors.IsOf(err, sdkerrors.ErrUnauthorized))
+}
+
+//nolint:funlen // this is complex test scenario and breaking it down is not helpful
+func TestKeeper_FreezeWhitelistMultiSend(t *testing.T) {
+	requireT := require.New(t)
+
+	testApp := simapp.New()
+	ctx := testApp.BaseApp.NewContext(false, tmproto.Header{})
+
+	ftKeeper := testApp.AssetFTKeeper
+	bankKeeper := testApp.BankKeeper
+
+	issuer1 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	issuer2 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	settings1 := types.IssueSettings{
+		Issuer:        issuer1,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(1000),
+		Features:      []types.Feature{types.Feature_freezing}, //nolint:nosnakecase
+	}
+
+	settings2 := types.IssueSettings{
+		Issuer:        issuer2,
+		Symbol:        "DEF",
+		Subunit:       "def",
+		Description:   "DEF Desc",
+		InitialAmount: sdk.NewInt(2000),
+		Features:      []types.Feature{types.Feature_whitelisting}, //nolint:nosnakecase
+	}
+
+	bondDenom := testApp.StakingKeeper.BondDenom(ctx)
+	// fund with the native coin
+	err := testApp.FundAccount(ctx, issuer1, sdk.NewCoins(sdk.NewCoin(bondDenom, sdk.NewInt(1000))))
+	requireT.NoError(err)
+
+	denom1, err := ftKeeper.Issue(ctx, settings1)
+	requireT.NoError(err)
+
+	denom2, err := ftKeeper.Issue(ctx, settings2)
+	requireT.NoError(err)
+
+	recipient1 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+	recipient2 := sdk.AccAddress(secp256k1.GenPrivKey().PubKey().Address())
+
+	// freeze denom1 partially on the recipient1
+	err = ftKeeper.Freeze(ctx, issuer1, recipient1, sdk.NewCoin(denom1, sdk.NewInt(10)))
+	requireT.NoError(err)
+
+	// whitelist denom2 partially on the recipient2
+	err = ftKeeper.SetWhitelistedBalance(ctx, issuer2, recipient2, sdk.NewCoin(denom2, sdk.NewInt(10)))
+	requireT.NoError(err)
+
+	// multi-send valid amount
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{
+			{Address: issuer1.String(), Coins: sdk.NewCoins(
+				sdk.NewCoin(denom1, sdk.NewInt(15)),
+				sdk.NewCoin(bondDenom, sdk.NewInt(20)),
+			)},
+			{Address: issuer2.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom2, sdk.NewInt(10)))},
+		},
+		[]banktypes.Output{
+			// the recipient1 has frozen balance so that amount can be received
+			{Address: recipient1.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom1, sdk.NewInt(15)))},
+			// the recipient2 has whitelisted balance so that is the max amount recipient2 can receive
+			{Address: recipient2.String(), Coins: sdk.NewCoins(
+				sdk.NewCoin(denom2, sdk.NewInt(10)),
+				sdk.NewCoin(bondDenom, sdk.NewInt(20)),
+			)},
+		})
+	requireT.NoError(err)
+
+	balance := bankKeeper.GetBalance(ctx, recipient1, denom1)
+	requireT.Equal(sdk.NewCoin(denom1, sdk.NewInt(15)).String(), balance.String())
+	balance = bankKeeper.GetBalance(ctx, recipient2, denom2)
+	requireT.Equal(sdk.NewCoin(denom2, sdk.NewInt(10)).String(), balance.String())
+	balance = bankKeeper.GetBalance(ctx, recipient2, bondDenom)
+	requireT.Equal(sdk.NewCoin(bondDenom, sdk.NewInt(20)).String(), balance.String())
+
+	// multi-send invalid frozen amount
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{
+			// we can't return 15 coins since 10 are frozen
+			{Address: recipient1.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom1, sdk.NewInt(15)))},
+			{Address: recipient2.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom2, sdk.NewInt(10)))},
+		},
+		[]banktypes.Output{
+			{Address: issuer1.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom1, sdk.NewInt(15)))},
+			{Address: issuer2.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom2, sdk.NewInt(10)))},
+		})
+	requireT.ErrorIs(sdkerrors.ErrInsufficientFunds, err)
+
+	// multi-send invalid whitelisted amount
+	err = bankKeeper.InputOutputCoins(ctx,
+		[]banktypes.Input{
+			{Address: issuer1.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom1, sdk.NewInt(15)))},
+			{Address: issuer2.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom2, sdk.NewInt(15)))},
+		},
+		[]banktypes.Output{
+			{Address: recipient1.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom1, sdk.NewInt(15)))},
+			// the recipient2 has whitelisted 10 so can't receive 15
+			{Address: recipient2.String(), Coins: sdk.NewCoins(sdk.NewCoin(denom2, sdk.NewInt(15)))},
+		})
+	requireT.ErrorIs(types.ErrWhitelistedLimitExceeded, err)
 }
 
 func TestKeeper_GetIssuerTokens(t *testing.T) {
@@ -1112,6 +1380,19 @@ type bankAssertion struct {
 	t   require.TestingT
 	bk  wbankkeeper.BaseKeeperWrapper
 	ctx sdk.Context
+}
+
+func (ba bankAssertion) assertCoinDistribution(denom string, dist map[*sdk.AccAddress]int64) {
+	requireT := require.New(ba.t)
+	total := int64(0)
+	for acc, expectedBalance := range dist {
+		total += expectedBalance
+		getBalance := ba.bk.GetBalance(ba.ctx, *acc, denom)
+		requireT.Equal(sdk.NewCoin(denom, sdk.NewInt(expectedBalance)).String(), getBalance.String())
+	}
+
+	totalSupply := ba.bk.GetSupply(ba.ctx, denom)
+	requireT.Equal(totalSupply.String(), sdk.NewCoin(denom, sdk.NewInt(total)).String())
 }
 
 func newBankAsserter(
