+++ x/asset/ft/types/genesis.go
@@ -1,6 +1,6 @@
 package types
 
-// DefaultGenesis returns the default FT genesis state.
+// DefaultGenesis returns the default Token genesis state.
 func DefaultGenesis() *GenesisState {
 	return &GenesisState{
 		Params: DefaultParams(),
